requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
flask-migrate
werkzeug
python-dotenv